CU-md1a77[Ready] - Add ACF to Home Page Template
