+++ header.php
@@ -96,9 +96,15 @@
         <nav class="navbar navbar-main navbar-expand-lg navbar-dark bg-dark" id="navbar-main">
             <div class="container">
                 <!-- Brand -->
-                <a class="navbar-brand" href="/">
-                    <img alt="Image placeholder" src="<?php echo get_theme_file_uri('/src/assets/img/brand/light-mono.svg'); ?>" id="navbar-logo">
-                </a>
+                <?php if (has_custom_logo()) :?>
+                    <?php 
+                        $custom_logo_id = get_theme_mod( 'custom_logo' );
+                        $image = wp_get_attachment_image_src( $custom_logo_id , 'full' );
+                    ?>
+                    <a class="navbar-brand" href="/">
+                        <img src="<?php echo $image[0]; ?>" alt="Logo" id="navbar-logo">
+                    </a>
+                <?php endif; ?>
                 <!-- Toggler -->
                 <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbar-main-collapse" aria-controls="navbar-main-collapse" aria-expanded="false" aria-label="Toggle navigation">
                     <span class="navbar-toggler-icon"></span>
@@ -122,7 +128,7 @@
                     <!-- Right navigation -->
                     <ul class="navbar-nav align-items-lg-center d-none d-lg-flex ml-lg-auto">
                         <!-- Auth -->
-                        <li class="nav-item">
+                        <!-- <li class="nav-item">
                             <a class="nav-link" href="../../pages/authentication/login-basic.html">
                                 Sign in
                             </a>
@@ -132,7 +138,7 @@
                                 <span class="btn-inner--icon"><i data-feather="shopping-bag"></i></span>
                                 <span class="btn-inner--text">Buy now</span>
                             </a>
-                        </li>
+                        </li> -->
                     </ul>
                     <!-- Mobile button -->
                     <!-- <div class="d-lg-none px-4 text-center">
