--- page-about.php
@@ -1,307 +1,0 @@
-<?php
-
-/**
- * Template Name: About Page
- *
- * @package HammerSport
- * @subpackage Quick Ui Kit
- * @since Quick Ui Kit
- */
-get_header(); ?>
-<main>
-    <!-- Main content -->
-    <section class="slice py-8 bg-dark bg-cover bg-size--cover"
-        style="background-image: url('<?php the_field('header_background'); ?>');">
-        <!-- Mask -->
-        <span class="mask bg-gradient-dark opacity-9"></span>
-        <!-- Container -->
-        <div class="container d-flex align-items-center text-center text-lg-left pt-5 pb-5 pt-lg-6">
-            <div class="col px-0">
-                <div class="row row-grid align-items-center">
-                    <div class="col-lg-8 text-center text-lg-left">
-                        <!-- Heading -->
-                        <h1 class="text-white mb-4">
-                            <?php the_field('headline'); ?>
-                        </h1>
-                        <!-- Text -->
-                        <p class="lead text-white opacity-8">
-                            <?php the_field('subheadline'); ?>
-                        </p>
-                        <!-- Buttons -->
-                        <?php
-                        if (get_field('header_call_to_action')) { ?>
-                        <div class="mt-5">
-                            <a href="#" class="btn btn-warning btn-lg btn-icon">
-                                <span class="btn-inner--icon">
-                                    <i data-feather="edit-3"></i>
-                                </span>
-                                <span class="btn-inner--text"><?php echo the_field('header_call_to_action') ?></span>
-                            </a>
-                        </div>
-                        <?php        } ?>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <!-- SVG separator -->
-        <div class="shape-container shape-line shape-position-bottom">
-            <svg width="2560px" height="100px" xmlns="http://www.w3.org/2000/svg"
-                xmlns:xlink="http://www.w3.org/1999/xlink" preserveAspectRatio="none" x="0px" y="0px"
-                viewBox="0 0 2560 100" style="enable-background:new 0 0 2560 100;" xml:space="preserve" class="">
-                <polygon points="2560 0 2560 100 0 100"></polygon>
-            </svg>
-        </div>
-    </section>
-    <section class="slice slice-lg">
-        <div class="container">
-            <span class="badge badge-primary badge-pill">Our vision</span>
-            <div class="row mt-4">
-                <div class="col-lg-6 pr-lg-5">
-                    <p class="h5 lh-180 mb-3">
-                        Our vision is to build the most efficient and intuitive UI stuff on the web in order to give
-                        developers some amazing possibilities to create high quality digital products with less time and
-                        money spent.
-                    </p>
-                </div>
-                <div class="col-lg-6">
-                    <p class="lead lh-180">
-                        She exposed painted fifteen are noisier mistake led waiting. Surprise not wandered speedily
-                        husbands although yet end. Are court tiled cease young built fat one man taken. We highest ye
-                        friends is exposed equally in. Ignorant had
-                        too strictly followed.
-                    </p>
-                    <p class="lead lh-180">
-                        Form face evening above years for i fruitful us creature void days. Upon upon fruitful us fill
-                        earth was set tree above yielding to air image cattle seed night lesser saying. Gathered, fourth
-                        wherein air.
-                    </p>
-                </div>
-            </div>
-        </div>
-    </section>
-    <section class="slice slice-lg">
-        <div class="container">
-            <div class="row">
-                <div class="col-lg-5 col-6">
-                    <img src="<?php echo get_theme_file_uri('/src/assets/img/theme/light/img-v-4.jpg'); ?>" alt="Image"
-                        class="rounded-lg img-fluid" />
-                    <div class="text-right mt-4">
-                        <img src="<?php echo get_theme_file_uri('/src/assets/img/theme/light/img-1-1000x1000.jpg'); ?>"
-                            alt="Image" class="rounded-lg img-fluid w-50" />
-                    </div>
-                </div>
-                <div class="col-lg-7 col-6">
-                    <img src="<?php echo get_theme_file_uri('/src/assets/img/theme/light/img-1-1000x800.jpg'); ?>"
-                        alt="Image" class="rounded-lg img-fluid w-75" />
-                    <div class="text-right mt-4">
-                        <img src="<?php echo get_theme_file_uri('/src/assets/img/theme/light/img-2-1000x800.jpg'); ?>"
-                            alt="Image" class="rounded-lg img-fluid" />
-                    </div>
-                </div>
-            </div>
-        </div>
-    </section>
-    <section class="slice slice-lg pt-10 bg-primary">
-        <!-- SVG separator -->
-        <div class="shape-container shape-line shape-position-top shape-orientation-inverse">
-            <svg width="2560px" height="100px" xmlns="http://www.w3.org/2000/svg"
-                xmlns:xlink="http://www.w3.org/1999/xlink" preserveAspectRatio="none" x="0px" y="0px"
-                viewBox="0 0 2560 100" style="enable-background:new 0 0 2560 100;" xml:space="preserve" class="">
-                <polygon points="2560 0 2560 100 0 100"></polygon>
-            </svg>
-        </div>
-        <!-- Container -->
-        <div class="container">
-            <div class="row row-grid">
-                <div class="col-lg-8">
-                    <!-- Badge -->
-                    <span class="badge badge-warning badge-pill">Made with Bootstrap</span>
-                    <!-- Title -->
-                    <h2 class="my-4 text-white">Websites, Dashboards and Web Apps</h2>
-                    <!-- Text -->
-                    <p class="lead text-white lh-190">
-                        Build one product from the ground up using our dedicated tools. A beautiful ecosystem built to
-                        sustain productivity and encourage developers to create modern and professional high-end
-                        products.
-                    </p>
-                </div>
-            </div>
-            <!-- Milestones -->
-            <div class="row mt-6">
-                <div class="col-lg-6">
-                    <div class="row">
-                        <div class="col-sm-4">
-                            <div class="card shadow-lg rounded-lg border-0 mb-sm-0">
-                                <div class="p-4 text-center text-sm-left">
-                                    <h3 class="mb-0">
-                                        <span class="counter" data-from="0" data-to="10" data-speed="3000"
-                                            data-refresh-interval="200"></span>
-                                        <span class="counter-extra">k</span>
-                                    </h3>
-                                    <p class="text-muted mb-0">
-                                        Customers
-                                    </p>
-                                </div>
-                            </div>
-                        </div>
-                        <div class="col-sm-4">
-                            <div class="card shadow-lg rounded-lg border-0 mb-sm-0">
-                                <div class="p-4 text-center text-sm-left">
-                                    <h3 class="mb-0">
-                                        <span class="counter" data-from="0" data-to="53" data-speed="3000"
-                                            data-refresh-interval="200"></span>
-                                        <span class="counter-extra">k</span>
-                                    </h3>
-                                    <p class="text-muted mb-0">
-                                        Downloads
-                                    </p>
-                                </div>
-                            </div>
-                        </div>
-                        <div class="col-sm-4">
-                            <div class="card shadow-lg rounded-lg border-0 mb-sm-0">
-                                <div class="p-4 text-center text-sm-left">
-                                    <h3 class="mb-0">
-                                        <span class="counter" data-from="0" data-to="98" data-speed="3000"
-                                            data-refresh-interval="200"></span>
-                                        <span class="counter-extra">%</span>
-                                    </h3>
-                                    <p class="text-muted mb-0">
-                                        Happy users
-                                    </p>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </section>
-    <section class="slice slice-lg pb-5">
-        <div class="container">
-            <!-- Section title -->
-            <div class="row mb-5 justify-content-center text-center">
-                <div class="col-lg-8 col-md-10">
-                    <h2 class=" mt-4">The amazing team</h2>
-                </div>
-            </div>
-            <div class="row">
-                <div class="col-lg-3 col-sm-6 mb-5">
-                    <div data-animate-hover="2">
-                        <div class="animate-this">
-                            <a href="#">
-                                <img alt="Image placeholder" class="img-fluid rounded shadow"
-                                    src="<?php echo get_theme_file_uri('/src/assets/img/theme/light/person-1.jpg'); ?>">
-                            </a>
-                        </div>
-                        <div class="mt-3">
-                            <h5 class="h6 mb-0">Heather Wright</h5>
-                            <p class="text-muted text-sm mb-0">Founded &amp; Chairman</p>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-lg-3 col-sm-6 mb-5">
-                    <div data-animate-hover="2">
-                        <div class="animate-this">
-                            <a href="#">
-                                <img alt="Image placeholder" class="img-fluid rounded shadow"
-                                    src="<?php echo get_theme_file_uri('/src/assets/img/theme/light/person-2.jpg') ?>">
-                            </a>
-                        </div>
-                        <div class="mt-3">
-                            <h5 class="h6 mb-0">Monroe Parker</h5>
-                            <p class="text-muted text-sm mb-0">Back End Developer</p>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-lg-3 col-sm-6 mb-5">
-                    <div data-animate-hover="2">
-                        <div class="animate-this">
-                            <a href="#">
-                                <img alt="Image placeholder" class="img-fluid rounded shadow"
-                                    src="<?php echo get_theme_file_uri('/src/assets/img/theme/light/person-3.jpg'); ?>">
-                            </a>
-                        </div>
-                        <div class="mt-3">
-                            <h5 class="h6 mb-0">John Sullivan</h5>
-                            <p class="text-muted text-sm mb-0">Front End Developer</p>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-lg-3 col-sm-6 mb-5">
-                    <div data-animate-hover="2">
-                        <div class="animate-this">
-                            <a href="#">
-                                <img alt="Image placeholder" class="img-fluid rounded shadow"
-                                    src="<?php echo get_theme_file_uri('/src/assets/img/theme/light/person-4.jpg'); ?>">
-                            </a>
-                        </div>
-                        <div class="mt-3">
-                            <h5 class="h6 mb-0">James Lewis</h5>
-                            <p class="text-muted text-sm mb-0">Vice Chairman</p>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="row">
-                <div class="col-lg-3 col-sm-6 mb-5 mb-sm-0">
-                    <div data-animate-hover="2">
-                        <div class="animate-this">
-                            <a href="#">
-                                <img alt="Image placeholder" class="img-fluid rounded shadow"
-                                    src="<?php echo get_theme_file_uri('/src/assets/img/theme/light/person-5.jpg'); ?>">
-                            </a>
-                        </div>
-                        <div class="mt-3">
-                            <h5 class="h6 mb-0">Danielle Levin</h5>
-                            <p class="text-muted text-sm mb-0">Sales Manager</p>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-lg-3 col-sm-6 mb-5 mb-sm-0">
-                    <div data-animate-hover="2">
-                        <div class="animate-this">
-                            <a href="#">
-                                <img alt="Image placeholder" class="img-fluid rounded shadow"
-                                    src="<?php echo get_theme_file_uri('/src/assets/img/theme/light/person-6.jpg'); ?>">
-                            </a>
-                        </div>
-                        <div class="mt-3">
-                            <h5 class="h6 mb-0">Martin Gray</h5>
-                            <p class="text-muted text-sm mb-0">UI/UX Designer</p>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-lg-3 col-sm-6 mb-5 mb-sm-0">
-                    <div data-animate-hover="2">
-                        <div class="animate-this">
-                            <a href="#">
-                                <img alt="Image placeholder" class="img-fluid rounded shadow"
-                                    src="<?php echo get_theme_file_uri('/src/assets/img/theme/light/person-7.jpg'); ?>">
-                            </a>
-                        </div>
-                        <div class="mt-3">
-                            <h5 class="h6 mb-0">George Squier</h5>
-                            <p class="text-muted text-sm mb-0">Marketing Executive</p>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-lg-3 col-sm-6 mb-0">
-                    <div data-animate-hover="2">
-                        <div class="animate-this">
-                            <a href="#">
-                                <img alt="Image placeholder" class="img-fluid rounded shadow"
-                                    src="<?php echo get_theme_file_uri('/src/assets/img/theme/light/person-8.jpg'); ?>">
-                            </a>
-                        </div>
-                        <div class="mt-3">
-                            <h5 class="h6 mb-0">Jesse Stevens</h5>
-                            <p class="text-muted text-sm mb-0">Ads Manager</p>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </section>
-</main>
-<?php get_footer(); ?>
